requirements.txt
# AI & NLP 라이브러리
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.3
openai==1.3.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4

# Discord 라이브러리
discord.py==2.3.2

# 데이터 처리
pandas==2.2.1
numpy==1.26.3

# 환경 및 시스템
python-dotenv==1.0.1

# 추가 선택적 라이브러리
asyncio==3.4.3
